requirement.txt
asgiref==3.3.0
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.1.3
django-environ==0.9.0
django-etc==1.3.1
django-import-export==2.5.0
django-templated-mail==1.1.1
djangorestframework==3.12.2
djangorestframework-simplejwt==4.5.0
djoser==2.0.3
et-xmlfile==1.1.0
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.7
Pillow==8.2.0
PyJWT==1.7.1
pytz==2020.4
PyYAML==5.4.1
sqlparse==0.4.1
tablib==3.0.0
xlrd==2.0.1
xlwt==1.3.0
